requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
opencv-python>=4.5.0
matplotlib>=3.4.0
pandas>=1.3.0
scikit-learn>=0.24.0
tqdm>=4.62.0
pillow>=8.3.0
tensorboard>=2.7.0 